@@ -37,17 +37,16 @@ def get_version(module='spyder_kernels'):
 
 REQUIREMENTS = [
     'cloudpickle',
-    'ipykernel>=6.16.1,<7',
-    'ipython>=7.31.1,<8',
-    'jupyter-client>=7.3.4,<8',
-    'packaging',
-    'pyzmq>=22.1.0',
+    'ipykernel>=6.29.3,<7',
+    'ipython>=8.12.2,<8.13; python_version=="3.8"',
+    'ipython>=8.13.0,<9,!=8.17.1; python_version>"3.8"',
+    'jupyter-client>=7.4.9,<9',
+    'pyzmq>=24.0.0',
     'wurlitzer>=1.0.3;platform_system!="Windows"',
     'pyxdg>=0.26;platform_system=="Linux"',
 ]
 
 TEST_REQUIREMENTS = [
-    'codecov',
     'cython',
     'dask[distributed]',
     'flaky',
@@ -61,6 +60,8 @@ TEST_REQUIREMENTS = [
     'xarray',
     'pillow',
     'django',
+    'h5py',
+    'pydicom'
 ]
 
 setup(
@@ -79,7 +80,7 @@ setup(
     install_requires=REQUIREMENTS,
     extras_require={'test': TEST_REQUIREMENTS},
     include_package_data=True,
-    python_requires='>=3.7',
+    python_requires='>=3.8',
     classifiers=[
         'Development Status :: 5 - Production/Stable',
         'Framework :: Jupyter',
@@ -88,10 +89,11 @@ setup(
         'License :: OSI Approved :: MIT License',
         'Operating System :: OS Independent',
         'Programming Language :: Python :: 3',
-        'Programming Language :: Python :: 3.7',
         'Programming Language :: Python :: 3.8',
         'Programming Language :: Python :: 3.9',
         'Programming Language :: Python :: 3.10',
+        'Programming Language :: Python :: 3.11',
+        'Programming Language :: Python :: 3.12',
         'Topic :: Software Development :: Interpreters',
     ]
 )
