@@ -23,9 +23,11 @@ from _thread import interrupt_main
 from ipykernel.zmqshell import ZMQInteractiveShell
 
 # Local imports
-import spyder_kernels
+from spyder_kernels.customize.namespace_manager import NamespaceManager
 from spyder_kernels.customize.spyderpdb import SpyderPdb
-from spyder_kernels.comms.frontendcomm import CommError
+from spyder_kernels.customize.code_runner import SpyderCodeRunner
+from spyder_kernels.comms.commbase import stacksummary_to_json
+from spyder_kernels.comms.decorators import comm_handler
 from spyder_kernels.utils.mpl import automatic_backend
 
 
@@ -45,22 +47,23 @@ class SpyderShell(ZMQInteractiveShell):
     ]
 
     def __init__(self, *args, **kwargs):
-        # Create _pdb_obj_stack before __init__
-        self._pdb_obj_stack = []
+        # Create _namespace_stack before __init__
+        self._namespace_stack = []
         self._request_pdb_stop = False
+        self.special = None
         self._pdb_conf = {}
         super(SpyderShell, self).__init__(*args, **kwargs)
         self._allow_kbdint = False
         self.register_debugger_sigint()
-
-        # Used for checking correct version by spyder
-        self._spyder_kernels_version = (
-            spyder_kernels.__version__,
-            sys.executable
-        )
+        self.update_gui_frontend = False
 
         # register post_execute
         self.events.register('post_execute', self.do_post_execute)
+
+    def init_magics(self):
+        """Init magics"""
+        super().init_magics()
+        self.register_magics(SpyderCodeRunner)
 
     def ask_exit(self):
         """Engage the exit actions."""
@@ -85,11 +88,31 @@ class SpyderShell(ZMQInteractiveShell):
         """Enable matplotlib."""
         if gui is None or gui.lower() == "auto":
             gui = automatic_backend()
-        gui, backend = super(SpyderShell, self).enable_matplotlib(gui)
-        try:
-            self.kernel.frontend_call(blocking=False).update_matplotlib_gui(gui)
-        except Exception:
-            pass
+
+        # Before activating the backend, restore to file default those
+        # InlineBackend settings that may have been set explicitly.
+        self.kernel.restore_rc_file_defaults()
+
+        enabled_gui, backend = super().enable_matplotlib(gui)
+
+        # This is necessary for IPython 8.24+, which returns None after
+        # enabling the Inline backend.
+        if enabled_gui is None and gui == "inline":
+            enabled_gui = "inline"
+        gui = enabled_gui
+
+        # To easily track the current interactive backend
+        if self.kernel.interactive_backend is None:
+            self.kernel.interactive_backend = gui if gui != "inline" else None
+
+        if self.update_gui_frontend:
+            try:
+                self.kernel.frontend_call(
+                    blocking=False
+                ).update_matplotlib_gui(gui)
+            except Exception:
+                pass
+
         return gui, backend
 
     # --- For Pdb namespace integration
@@ -109,31 +132,21 @@ class SpyderShell(ZMQInteractiveShell):
                 if self.pdb_session:
                     setattr(self.pdb_session, key, pdb_conf[key])
 
-    def get_local_scope(self, stack_depth):
-        """Get local scope at given frame depth."""
-        frame = sys._getframe(stack_depth + 1)
-        if self._pdb_frame is frame:
-            # Avoid calling f_locals on _pdb_frame
-            return self.pdb_session.curframe_locals
-        else:
-            return frame.f_locals
-
-    def get_global_scope(self, stack_depth):
-        """Get global scope at given frame depth."""
-        frame = sys._getframe(stack_depth + 1)
-        return frame.f_globals
-
     def is_debugging(self):
         """
         Check if we are currently debugging.
         """
-        return bool(self._pdb_frame)
+        for session in self._namespace_stack[::-1]:
+            if isinstance(session, SpyderPdb) and session.curframe is not None:
+                return True
+        return False
 
     @property
     def pdb_session(self):
         """Get current pdb session."""
-        if len(self._pdb_obj_stack) > 0:
-            return self._pdb_obj_stack[-1]
+        for session in self._namespace_stack[::-1]:
+            if isinstance(session, SpyderPdb):
+                return session
         return None
 
     def add_pdb_session(self, pdb_obj):
@@ -141,33 +154,60 @@ class SpyderShell(ZMQInteractiveShell):
         if self.pdb_session == pdb_obj:
             # Already added
             return
-        self._pdb_obj_stack.append(pdb_obj)
+        self._namespace_stack.append(pdb_obj)
 
         # Set config to pdb obj
         self.set_pdb_configuration(self._pdb_conf)
-
-        try:
-            self.kernel.frontend_call(blocking=False).set_debug_state(
-                len(self._pdb_obj_stack))
-        except (CommError, TimeoutError):
-            logger.debug("Could not send debugging state to the frontend.")
 
     def remove_pdb_session(self, pdb_obj):
         """Remove a pdb object from the stack."""
         if self.pdb_session != pdb_obj:
             # Already removed
             return
-        self._pdb_obj_stack.pop()
+        self._namespace_stack.pop()
 
         if self.pdb_session:
             # Set config to newly active pdb obj
             self.set_pdb_configuration(self._pdb_conf)
 
-        try:
-            self.kernel.frontend_call(blocking=False).set_debug_state(
-                len(self._pdb_obj_stack))
-        except (CommError, TimeoutError):
-            logger.debug("Could not send debugging state to the frontend.")
+    def add_namespace_manager(self, ns_manager):
+        """Add namespace manager to stack."""
+        self._namespace_stack.append(ns_manager)
+
+    def remove_namespace_manager(self, ns_manager):
+        """Remove namespace manager."""
+        if self._namespace_stack[-1] != ns_manager:
+            logger.debug("The namespace stack is inconsistent.")
+            return
+        self._namespace_stack.pop()
+
+    def get_local_scope(self, stack_depth):
+        """
+        Get local scope at a given frame depth.
+
+        Needed for magics that use "needs_local_scope" such as timeit
+        """
+        frame = sys._getframe(stack_depth + 1)
+        return self.context_locals(frame)
+
+    def context_locals(self, frame=None):
+        """
+        Get context locals.
+
+        If frame is not None, make sure frame.f_locals is not registered in a
+        debugger and return frame.f_locals
+        """
+        for session in self._namespace_stack[::-1]:
+            if isinstance(session, SpyderPdb) and session.curframe is not None:
+                if frame is None or frame == session.curframe:
+                    return session.curframe_locals
+            elif frame is None and isinstance(session, NamespaceManager):
+                return session.ns_locals
+
+        if frame is not None:
+            return frame.f_locals
+
+        return None
 
     @property
     def _pdb_frame(self):
@@ -176,28 +216,56 @@ class SpyderShell(ZMQInteractiveShell):
             return self.pdb_session.curframe
 
     @property
-    def _pdb_locals(self):
-        """
-        Return current Pdb frame locals if available. Otherwise
-        return an empty dictionary
-        """
-        if self._pdb_frame is not None:
-            return self.pdb_session.curframe_locals
-        else:
-            return {}
-
-    @property
     def user_ns(self):
         """Get the current namespace."""
-        if self._pdb_frame is not None:
-            return self._pdb_frame.f_globals
-        else:
-            return self.__user_ns
+        for session in self._namespace_stack[::-1]:
+            if isinstance(session, SpyderPdb) and session.curframe is not None:
+                # Return first debugging namespace
+                return session.curframe.f_globals
+            elif isinstance(session, NamespaceManager):
+                return session.ns_globals
+
+        return self.__user_ns
 
     @user_ns.setter
     def user_ns(self, namespace):
         """Set user_ns."""
         self.__user_ns = namespace
+
+    def _get_current_namespace(self, with_magics=False, frame=None):
+        """Return a copy of the current namespace."""
+        if frame is not None:
+            ns = frame.f_globals.copy()
+            ns.update(self.context_locals(frame))
+            return ns
+
+        ns = {}
+        ns.update(self.user_ns)
+        context_locals = self.context_locals()
+        if context_locals:
+            ns.update(context_locals)
+
+        # Add magics to ns so we can show help about them on the Help
+        # plugin
+        if with_magics:
+            line_magics = self.magics_manager.magics['line']
+            cell_magics = self.magics_manager.magics['cell']
+            ns.update(line_magics)
+            ns.update(cell_magics)
+
+        return ns
+
+    def _get_reference_namespace(self, name):
+        """
+        Return namespace where reference name is defined
+
+        It returns the user namespace if name has not yet been defined.
+        """
+        lcls = self.context_locals()
+        if lcls and name in lcls:
+
+            return lcls
+        return self.user_ns
 
     def showtraceback(self, exc_tuple=None, filename=None, tb_offset=None,
                       exception_only=False, running_compiled_code=False):
@@ -208,11 +276,9 @@ class SpyderShell(ZMQInteractiveShell):
         if not exception_only:
             try:
                 etype, value, tb = self._get_exc_info(exc_tuple)
-                stack = traceback.extract_tb(tb.tb_next)
-                for f_summary, f in zip(
-                        stack, traceback.walk_tb(tb.tb_next)):
-                    f_summary.locals = self.kernel.get_namespace_view(
-                        frame=f[0])
+                etype = etype.__name__
+                value = value.args
+                stack = stacksummary_to_json(traceback.extract_tb(tb.tb_next))
                 self.kernel.frontend_call(blocking=False).show_traceback(
                     etype, value, stack)
             except Exception:
@@ -222,6 +288,7 @@ class SpyderShell(ZMQInteractiveShell):
         """Register sigint handler."""
         signal.signal(signal.SIGINT, self.spyderkernel_sigint_handler)
 
+    @comm_handler
     def raise_interrupt_signal(self):
         """Raise interrupt signal."""
         if os.name == "nt":
@@ -233,8 +300,15 @@ class SpyderShell(ZMQInteractiveShell):
                 self.kernel.log.error(
                     "Interrupt message not supported on Windows")
         else:
-            self.kernel._send_interupt_children()
+            # This is necessary to make the call below work for IPykernel
+            # versions equal or less than 6.21.2 and greater than it.
+            # See ipython/ipykernel#1101
+            if hasattr(self.kernel, '_send_interupt_children'):
+                self.kernel._send_interupt_children()
+            else:
+                self.kernel._send_interrupt_children()
 
+    @comm_handler
     def request_pdb_stop(self):
         """Request pdb to stop at the next possible position."""
         pdb_session = self.pdb_session
@@ -293,6 +367,7 @@ class SpyderShell(ZMQInteractiveShell):
         except KeyboardInterrupt:
             self.showtraceback()
 
+    @comm_handler
     def pdb_input_reply(self, line, echo_stack_entry=True):
         """Get a pdb command from the frontend."""
         debugger = self.pdb_session
@@ -308,3 +383,4 @@ class SpyderShell(ZMQInteractiveShell):
         # Flush C standard streams.
         sys.__stderr__.flush()
         sys.__stdout__.flush()
+        self.kernel.publish_state()
