@@ -8,7 +8,8 @@
 Class that handles communications between Spyder kernel and frontend.
 
 Comms transmit data in a list of buffers, and in a json-able dictionnary.
-Here, we only support a buffer list with a single element.
+Here, we only support json to avoid issues of compatibility between Python
+versions. In the abstraction below, buffers is used to send bytes.
 
 The messages exchanged have the following msg_dict:
 
@@ -18,8 +19,6 @@ The messages exchanged have the following msg_dict:
         'content': content,
     }
     ```
-
-The buffer is generated by cloudpickle using `PICKLE_PROTOCOL = 2`.
 
 To simplify the usage of messaging, we use a higher level function calling
 mechanism:
@@ -35,34 +34,31 @@ The messages exchanged are:
             'call_name': The name of the function to be called,
             'call_id': uuid to match the request to a potential reply,
             'settings': A dictionnary of settings,
-            }
-        - The buffer encodes a dictionnary {
             'call_args': The function args,
             'call_kwargs': The function kwargs,
-            }
+            'buffered_args': The args index that are in the buffers,
+            'buffered_kwargs': the kwargs keys that are in the buffers
+          }
+        - The buffer contains any bytes in the arguments
     - If the 'settings' has `'blocking' =  True`, a reply is sent.
       (spyder_msg_type = 'remote_call_reply'):
-        - The buffer contains the return value of the function.
         - The 'content' is a dict with: {
-                'is_error': a boolean indicating if the return value is an
-                            exception to be raised.
-                'call_id': The uuid from above,
-                'call_name': The function name (mostly for debugging)
-                }
+            'is_error': a boolean indicating if the return value is an
+                        exception to be raised.
+            'call_id': The uuid from above,
+            'call_name': The function name (mostly for debugging),
+            'call_return_value': The return value of the function
+           }
+        - The buffer contains the return value if it is bytes
 """
-from __future__ import print_function
-
-import cloudpickle
-import pickle
 import logging
 import sys
 import uuid
 import traceback
+import builtins
 
 
 logger = logging.getLogger(__name__)
-
-DEFAULT_PICKLE_PROTOCOL = 4
 
 # Max timeout (in secs) for blocking calls
 TIMEOUT = 3
@@ -72,12 +68,64 @@ class CommError(RuntimeError):
     pass
 
 
+def stacksummary_to_json(stack):
+    """StackSummary to json."""
+    return [
+        {
+            "filename": frame.filename,
+            "lineno": frame.lineno,
+            "name": frame.name,
+            "line": frame.line
+        }
+        for frame in stack
+    ]
+
+
+def staksummary_from_json(stack):
+    """StackSummary from json."""
+    traceback.StackSummary.from_list([
+        (
+            frame["filename"],
+            frame["lineno"],
+            frame["name"],
+            frame["line"]
+        )
+        for frame in stack
+    ])
+
+
 class CommsErrorWrapper():
     def __init__(self, call_name, call_id):
         self.call_name = call_name
         self.call_id = call_id
         self.etype, self.error, tb = sys.exc_info()
         self.tb = traceback.extract_tb(tb)
+
+    def to_json(self):
+        """Create JSON representation."""
+        return {
+            "call_name": self.call_name,
+            "call_id": self.call_id,
+            "etype": self.etype.__name__,
+            "args": self.error.args,
+            "tb": stacksummary_to_json(self.tb)
+        }
+
+    @classmethod
+    def from_json(cls, json_data):
+        """Get a CommsErrorWrapper from a JSON representation."""
+        instance = cls.__new__(cls)
+        instance.call_name = json_data["call_name"]
+        instance.call_id = json_data["call_id"]
+        etype = json_data["etype"]
+        instance.etype = getattr(
+            builtins,
+            etype,
+            type(etype, (Exception,), {})
+        )
+        instance.error = instance.etype(*json_data["args"])
+        instance.tb = staksummary_from_json(json_data["tb"])
+        return instance
 
     def raise_error(self):
         """
@@ -206,8 +254,9 @@ class CommBase:
         return RemoteCallFactory(self, comm_id, callback, **settings)
 
     # ---- Private -----
-    def _send_message(self, spyder_msg_type, content=None, data=None,
-                      comm_id=None):
+    def _send_message(
+        self, spyder_msg_type, content=None, comm_id=None, buffers=None
+    ):
         """
         Publish custom messages to the other side.
 
@@ -217,11 +266,10 @@ class CommBase:
             The spyder message type
         content: dict
             The (JSONable) content of the message
-        data: any
-            Any object that is serializable by cloudpickle (should be most
-            things). Will arrive as cloudpickled bytes in `.buffers[0]`.
         comm_id: int
             the comm to send to. If None sends to all comms.
+        buffers: list(bytes)
+            a list of bytes to send.
         """
         if not self.is_open(comm_id):
             raise CommError("The comm is not connected.")
@@ -230,17 +278,9 @@ class CommBase:
             msg_dict = {
                 'spyder_msg_type': spyder_msg_type,
                 'content': content,
-                'pickle_protocol': self._comms[comm_id]['pickle_protocol'],
-                'python_version': sys.version,
-                }
-            buffers = [cloudpickle.dumps(
-                data, protocol=self._comms[comm_id]['pickle_protocol'])]
-            self._comms[comm_id]['comm'].send(msg_dict, buffers=buffers)
+            }
 
-    def _set_pickle_protocol(self, protocol):
-        """Set the pickle protocol used to send data."""
-        protocol = min(protocol, pickle.HIGHEST_PROTOCOL)
-        self._comms[self.calling_comm_id]['pickle_protocol'] = protocol
+            self._comms[comm_id]['comm'].send(msg_dict, buffers=buffers)
 
     @property
     def _comm_name(self):
@@ -258,9 +298,8 @@ class CommBase:
         message_id : str
             The identifier for the message
         handler : callback
-            A function to handle the message. This is called with 3 arguments:
+            A function to handle the message. This is called with:
                 - msg_dict: A dictionary with message information.
-                - buffer: The data transmitted in the buffer
             Pass None to unregister the message_id
         """
         if handler is None:
@@ -277,9 +316,8 @@ class CommBase:
         comm.on_close(self._comm_close)
         self._comms[comm.comm_id] = {
             'comm': comm,
-            'pickle_protocol': DEFAULT_PICKLE_PROTOCOL,
             'status': 'opening',
-            }
+        }
 
     def _comm_close(self, msg):
         """Close comm."""
@@ -294,36 +332,35 @@ class CommBase:
 
         # Get message dict
         msg_dict = msg['content']['data']
-
-        # Load the buffer. Only one is supported.
-        try:
-            buffer = cloudpickle.loads(msg['buffers'][0])
-        except Exception as e:
-            logger.debug(
-                "Exception in cloudpickle.loads : %s" % str(e))
-            buffer = CommsErrorWrapper(
-                msg_dict['content']['call_name'],
-                msg_dict['content']['call_id'])
-
-            msg_dict['content']['is_error'] = True
-
         spyder_msg_type = msg_dict['spyder_msg_type']
+        buffers = msg['buffers']
 
         if spyder_msg_type in self._message_handlers:
-            self._message_handlers[spyder_msg_type](
-                msg_dict, buffer)
+            self._message_handlers[spyder_msg_type](msg_dict, buffers)
         else:
             logger.debug("No such spyder message type: %s" % spyder_msg_type)
 
-    def _handle_remote_call(self, msg, buffer):
+    def _handle_remote_call(self, msg, buffers):
         """Handle a remote call."""
         msg_dict = msg['content']
         self.on_incoming_call(msg_dict)
         try:
+            # read buffers
+            args = msg_dict['call_args']
+            kwargs = msg_dict['call_kwargs']
+
+            if buffers:
+                for idx in msg_dict['buffered_args']:
+                    args[idx] = buffers.pop(0)
+                for name in msg_dict['buffered_kwargs']:
+                    kwargs[name] = buffers.pop(0)
+                assert len(buffers) == 0
+
             return_value = self._remote_callback(
-                    msg_dict['call_name'],
-                    buffer['call_args'],
-                    buffer['call_kwargs'])
+                msg_dict['call_name'],
+                args,
+                kwargs
+            )
             self._set_call_return_value(msg_dict, return_value)
         except Exception:
             exc_infos = CommsErrorWrapper(
@@ -338,7 +375,7 @@ class CommBase:
 
         raise CommError("No such spyder call type: %s" % call_name)
 
-    def _set_call_return_value(self, call_dict, data, is_error=False):
+    def _set_call_return_value(self, call_dict, return_value, is_error=False):
         """
         A remote call has just been processed.
 
@@ -348,21 +385,34 @@ class CommBase:
 
         display_error = ('display_error' in settings and
                          settings['display_error'])
-        if is_error and display_error:
-            data.print_error()
+        if is_error:
+            if display_error:
+                return_value.print_error()
+            return_value = return_value.to_json()
 
         send_reply = 'send_reply' in settings and settings['send_reply']
         if not send_reply:
             # Nothing to send back
             return
+
+        buffers = None
+        if isinstance(return_value, bytes):
+            buffers = [return_value]
+            return_value = None
+
         content = {
             'is_error': is_error,
             'call_id': call_dict['call_id'],
-            'call_name': call_dict['call_name']
+            'call_name': call_dict['call_name'],
+            'call_return_value': return_value
         }
 
-        self._send_message('remote_call_reply', content=content, data=data,
-                           comm_id=self.calling_comm_id)
+        self._send_message(
+            'remote_call_reply',
+            content=content,
+            comm_id=self.calling_comm_id,
+            buffers=buffers
+        )
 
     def _register_call(self, call_dict, callback=None):
         """
@@ -376,20 +426,18 @@ class CommBase:
 
     def on_outgoing_call(self, call_dict):
         """A message is about to be sent"""
-        call_dict["pickle_highest_protocol"] = pickle.HIGHEST_PROTOCOL
         return call_dict
 
     def on_incoming_call(self, call_dict):
         """A call was received"""
-        if "pickle_highest_protocol" in call_dict:
-            self._set_pickle_protocol(call_dict["pickle_highest_protocol"])
+        pass
 
-    def _send_call(self, call_dict, call_data, comm_id):
+    def _send_call(self, call_dict, comm_id, buffers=None):
         """Send call."""
         call_dict = self.on_outgoing_call(call_dict)
         self._send_message(
-            'remote_call', content=call_dict, data=call_data,
-            comm_id=comm_id)
+            'remote_call', content=call_dict, comm_id=comm_id, buffers=buffers
+        )
 
     def _get_call_return_value(self, call_dict, comm_id):
         """
@@ -401,7 +449,6 @@ class CommBase:
         settings = call_dict['settings']
 
         blocking = 'blocking' in settings and settings['blocking']
-
         if not blocking:
             return
 
@@ -416,12 +463,12 @@ class CommBase:
 
         self._wait_reply(comm_id, call_id, call_name, timeout)
 
-        reply = self._reply_inbox.pop(call_id)
+        content = self._reply_inbox.pop(call_id)
+        return_value = content['call_return_value']
 
-        if reply['is_error']:
-            return self._sync_error(reply['value'])
-
-        return reply['value']
+        if content['is_error']:
+            return self._sync_error(return_value)
+        return return_value
 
     def _wait_reply(self, comm_id, call_id, call_name, timeout):
         """
@@ -429,7 +476,7 @@ class CommBase:
         """
         raise NotImplementedError
 
-    def _handle_remote_call_reply(self, msg_dict, buffer):
+    def _handle_remote_call_reply(self, msg_dict, buffers):
         """
         A blocking call received a reply.
         """
@@ -437,11 +484,20 @@ class CommBase:
         call_id = content['call_id']
         call_name = content['call_name']
         is_error = content['is_error']
+        return_value = content['call_return_value']
+
+        # Prepare return value
+        if is_error:
+            return_value = CommsErrorWrapper.from_json(return_value)
+        elif buffers:
+            assert len(buffers) == 1
+            return_value = buffers[0]
+        content['call_return_value'] = return_value
 
         # Unexpected reply
         if call_id not in self._reply_waitlist:
             if is_error:
-                return self._async_error(buffer)
+                return self._async_error(return_value)
             else:
                 logger.debug('Got an unexpected reply {}, id:{}'.format(
                     call_name, call_id))
@@ -451,19 +507,15 @@ class CommBase:
 
         # Async error
         if is_error and not blocking:
-            return self._async_error(buffer)
+            return self._async_error(return_value)
 
         # Callback
         if callback is not None and not is_error:
-            callback(buffer)
+            callback(return_value)
 
         # Blocking inbox
         if blocking:
-            self._reply_inbox[call_id] = {
-                    'is_error': is_error,
-                    'value': buffer,
-                    'content': content
-                    }
+            self._reply_inbox[call_id] = content
 
     def _async_error(self, error_wrapper):
         """
@@ -513,21 +565,40 @@ class RemoteCall():
         """
         Transmit the call to the other side of the tunnel.
 
-        The args and kwargs have to be picklable.
+        The args and kwargs have to be JSON-serializable or bytes.
         """
         blocking = 'blocking' in self._settings and self._settings['blocking']
         self._settings['send_reply'] = blocking or self._callback is not None
+
+        # The call will be serialized with json. The bytes are sent separately.
+        buffers = []
+        buffered_args = []
+        buffered_kwargs = []
+        args = list(args)
+
+        for i, arg in enumerate(args):
+            if isinstance(arg, bytes):
+                buffers.append(arg)
+                buffered_args.append(i)
+                args[i] = None
+
+        for name in kwargs:
+            arg = kwargs[name]
+            if isinstance(arg, bytes):
+                buffers.append(arg)
+                buffered_kwargs.append(name)
+                kwargs[name] = None
 
         call_id = uuid.uuid4().hex
         call_dict = {
             'call_name': self._name,
             'call_id': call_id,
             'settings': self._settings,
-            }
-        call_data = {
             'call_args': args,
             'call_kwargs': kwargs,
-            }
+            'buffered_args': buffered_args,
+            'buffered_kwargs': buffered_kwargs
+        }
 
         if not self._comms_wrapper.is_open(self._comm_id):
             # Only an error if the call is blocking.
@@ -536,6 +607,6 @@ class RemoteCall():
             logger.debug("Call to unconnected comm: %s" % self._name)
             return
         self._comms_wrapper._register_call(call_dict, self._callback)
-        self._comms_wrapper._send_call(call_dict, call_data, self._comm_id)
+        self._comms_wrapper._send_call(call_dict, self._comm_id, buffers)
         return self._comms_wrapper._get_call_return_value(
             call_dict, self._comm_id)
