@@ -17,13 +17,10 @@ import sys
 import site
 
 # Third-party imports
-from packaging.version import parse
 from traitlets import DottedObjectName
 
 # Local imports
 from spyder_kernels.utils.misc import is_module_installed
-from spyder_kernels.utils.mpl import (
-    MPL_BACKENDS_FROM_SPYDER, INLINE_FIGURE_FORMATS)
 
 
 def import_spydercustomize():
@@ -48,24 +45,6 @@ def import_spydercustomize():
         sys.path.remove(customize_dir)
     except ValueError:
         pass
-
-
-def sympy_config(mpl_backend):
-    """Sympy configuration"""
-    if mpl_backend is not None:
-        lines = """
-from sympy.interactive import init_session
-init_session()
-%matplotlib {0}
-""".format(mpl_backend)
-    else:
-        lines = """
-from sympy.interactive import init_session
-init_session()
-"""
-
-    return lines
-
 
 def kernel_config():
     """Create a config object with IPython kernel options."""
@@ -105,14 +84,6 @@ def kernel_config():
     clear_argv = "import sys; sys.argv = ['']; del sys"
     spy_cfg.IPKernelApp.exec_lines = [clear_argv]
 
-    # Set our runfile in builtins here to prevent other packages shadowing it.
-    # This started to be a problem since IPykernel 6.3.0.
-    spy_cfg.IPKernelApp.exec_lines.append(
-        "import builtins; "
-        "builtins.runfile = builtins.spyder_runfile; "
-        "del builtins.spyder_runfile; del builtins"
-    )
-
     # Prevent other libraries to change the breakpoint builtin.
     # This started to be a problem since IPykernel 6.3.0.
     if sys.version_info[0:2] >= (3, 7):
@@ -122,100 +93,8 @@ def kernel_config():
             "del sys; del pdb"
         )
 
-    # Run lines of code at startup
-    run_lines_o = os.environ.get('SPY_RUN_LINES_O')
-    if run_lines_o is not None:
-        spy_cfg.IPKernelApp.exec_lines += (
-            [x.strip() for x in run_lines_o.split(';')]
-        )
-
-    # Load %autoreload magic
-    spy_cfg.IPKernelApp.exec_lines.append(
-        "get_ipython().kernel._load_autoreload_magic()")
-
-    # Load wurlitzer extension
-    spy_cfg.IPKernelApp.exec_lines.append(
-        "get_ipython().kernel._load_wurlitzer()")
-
-    # Default inline backend configuration.
-    # This is useful to have when people doesn't
-    # use our config system to configure the
-    # inline backend but want to use
-    # '%matplotlib inline' at runtime
-    spy_cfg.InlineBackend.rc = {
-        # The typical default figure size is too large for inline use,
-        # so we shrink the figure size to 6x4, and tweak fonts to
-        # make that fit.
-        'figure.figsize': (6.0, 4.0),
-        # 72 dpi matches SVG/qtconsole.
-        # This only affects PNG export, as SVG has no dpi setting.
-        'figure.dpi': 72,
-        # 12pt labels get cutoff on 6x4 logplots, so use 10pt.
-        'font.size': 10,
-        # 10pt still needs a little more room on the xlabel
-        'figure.subplot.bottom': .125,
-        # Play nicely with any background color.
-        'figure.facecolor': 'white',
-        'figure.edgecolor': 'white'
-    }
-
-    # Pylab configuration
-    mpl_backend = None
     if is_module_installed('matplotlib'):
-        # Set Matplotlib backend with Spyder options
-        pylab_o = os.environ.get('SPY_PYLAB_O')
-        backend_o = os.environ.get('SPY_BACKEND_O')
-        if pylab_o == 'True' and backend_o is not None:
-            mpl_backend = MPL_BACKENDS_FROM_SPYDER[backend_o]
-            # Inline backend configuration
-            if mpl_backend == 'inline':
-                # Figure format
-                format_o = os.environ.get('SPY_FORMAT_O')
-                formats = INLINE_FIGURE_FORMATS
-                if format_o is not None:
-                    spy_cfg.InlineBackend.figure_format = formats[format_o]
-
-                # Resolution
-                resolution_o = os.environ.get('SPY_RESOLUTION_O')
-                if resolution_o is not None:
-                    spy_cfg.InlineBackend.rc['figure.dpi'] = float(
-                        resolution_o)
-
-                # Figure size
-                width_o = float(os.environ.get('SPY_WIDTH_O'))
-                height_o = float(os.environ.get('SPY_HEIGHT_O'))
-                if width_o is not None and height_o is not None:
-                    spy_cfg.InlineBackend.rc['figure.figsize'] = (width_o,
-                                                                  height_o)
-
-                # Print figure kwargs
-                bbox_inches_o = os.environ.get('SPY_BBOX_INCHES_O')
-                bbox_inches = 'tight' if bbox_inches_o == 'True' else None
-                spy_cfg.InlineBackend.print_figure_kwargs.update(
-                    {'bbox_inches': bbox_inches})
-        else:
-            # Set Matplotlib backend to inline for external kernels.
-            # Fixes issue 108
-            mpl_backend = 'inline'
-
-        # Automatically load Pylab and Numpy, or only set Matplotlib
-        # backend
-        autoload_pylab_o = os.environ.get('SPY_AUTOLOAD_PYLAB_O') == 'True'
-        command = "get_ipython().kernel._set_mpl_backend('{0}', {1})"
-        spy_cfg.IPKernelApp.exec_lines.append(
-            command.format(mpl_backend, autoload_pylab_o))
-
-    # Enable Cython magic
-    run_cython = os.environ.get('SPY_RUN_CYTHON') == 'True'
-    if run_cython and is_module_installed('Cython'):
-        spy_cfg.IPKernelApp.exec_lines.append('%reload_ext Cython')
-
-    # Run a file at startup
-    use_file_o = os.environ.get('SPY_USE_FILE_O')
-    run_file_o = os.environ.get('SPY_RUN_FILE_O')
-    if use_file_o == 'True' and run_file_o is not None:
-        if osp.exists(run_file_o):
-            spy_cfg.IPKernelApp.file_to_run = run_file_o
+        spy_cfg.IPKernelApp.matplotlib = "inline"
 
     # Autocall
     autocall_o = os.environ.get('SPY_AUTOCALL_O')
@@ -228,12 +107,6 @@ def kernel_config():
     # Greedy completer
     greedy_o = os.environ.get('SPY_GREEDY_O') == 'True'
     spy_cfg.IPCompleter.greedy = greedy_o
-
-    # Sympy loading
-    sympy_o = os.environ.get('SPY_SYMPY_O') == 'True'
-    if sympy_o and is_module_installed('sympy'):
-        lines = sympy_config(mpl_backend)
-        spy_cfg.IPKernelApp.exec_lines.append(lines)
 
     # Disable the new mechanism to capture and forward low-level output
     # in IPykernel 6. For that we have Wurlitzer.
@@ -255,7 +128,7 @@ def varexp(line):
     ip = get_ipython()       #analysis:ignore
     funcname, name = line.split()
     try:
-        data = ip.kernel._get_current_namespace()[name]
+        data = ip._get_current_namespace()[name]
     except KeyError:
         print("\033[41mVariable %s does not exist in current namespace.\033[0m" % name)
     plotlib = os.environ.get('SPY_VAREXP_PLOTLIB')
