@@ -11,30 +11,13 @@
 # Spyder consoles sitecustomize
 #
 
-import ast
-import bdb
-import builtins
-import io
 import logging
 import os
 import pdb
-import shlex
 import sys
-import time
 import warnings
 
-from IPython.core.getipython import get_ipython
-from IPython.core.inputtransformer2 import (
-    TransformerManager, leading_indent, leading_empty_lines)
-
-from spyder_kernels.comms.frontendcomm import frontend_request
-from spyder_kernels.customize.namespace_manager import NamespaceManager
-from spyder_kernels.customize.spyderpdb import SpyderPdb, get_new_debugger
-from spyder_kernels.customize.umr import UserModuleReloader
-from spyder_kernels.customize.utils import capture_last_Expr, canonic
-
-
-logger = logging.getLogger(__name__)
+from spyder_kernels.customize.spyderpdb import SpyderPdb
 
 
 # =============================================================================
@@ -45,15 +28,11 @@ logger = logging.getLogger(__name__)
 if not hasattr(sys, 'argv'):
     sys.argv = ['']
 
-
 # =============================================================================
 # Main constants
 # =============================================================================
 IS_EXT_INTERPRETER = os.environ.get('SPY_EXTERNAL_INTERPRETER') == "True"
 HIDE_CMD_WINDOWS = os.environ.get('SPY_HIDE_CMD') == "True"
-SHOW_INVALID_SYNTAX_MSG = True
-SHOW_GLOBAL_MSG = True
-
 
 # =============================================================================
 # Prevent subprocess.Popen calls to create visible console windows on Windows.
@@ -70,7 +49,6 @@ if os.name == 'nt' and HIDE_CMD_WINDOWS:
 
     subprocess.Popen = SubprocessPopen
 
-
 # =============================================================================
 # Importing user's sitecustomize
 # =============================================================================
@@ -78,7 +56,6 @@ try:
     import sitecustomize  #analysis:ignore
 except Exception:
     pass
-
 
 # =============================================================================
 # Set PyQt API to #2
@@ -275,450 +252,6 @@ os.get_terminal_size = _patched_get_terminal_size
 # Pdb adjustments
 # =============================================================================
 pdb.Pdb = SpyderPdb
-
-# =============================================================================
-# User module reloader
-# =============================================================================
-__umr__ = UserModuleReloader(namelist=os.environ.get("SPY_UMR_NAMELIST", None))
-
-
-# =============================================================================
-# Handle Post Mortem Debugging and Traceback Linkage to Spyder
-# =============================================================================
-def post_mortem_excepthook(type, value, tb):
-    """
-    For post mortem exception handling, print a banner and enable post
-    mortem debugging.
-    """
-    ipython_shell = get_ipython()
-    ipython_shell.showtraceback((type, value, tb))
-    p = pdb.Pdb(ipython_shell.colors)
-
-    if not type == SyntaxError:
-        # wait for stderr to print (stderr.flush does not work in this case)
-        time.sleep(0.1)
-        print('*' * 40)
-        print('Entering post mortem debugging...')
-        print('*' * 40)
-        #  add ability to move between frames
-        p.reset()
-        frame = tb.tb_next.tb_frame
-        # wait for stdout to print
-        time.sleep(0.1)
-        p.interaction(frame, tb)
-
-
-# ==============================================================================
-# runfile and debugfile commands
-# ==============================================================================
-def get_current_file_name():
-    """Get the current file name."""
-    try:
-        return frontend_request(blocking=True).current_filename()
-    except Exception:
-        print("This command failed to be executed because an error occurred"
-              " while trying to get the current file name from Spyder's"
-              " editor. The error was:\n\n")
-        get_ipython().showtraceback(exception_only=True)
-        return None
-
-
-def count_leading_empty_lines(cell):
-    """Count the number of leading empty cells."""
-    lines = cell.splitlines(keepends=True)
-    if not lines:
-        return 0
-    for i, line in enumerate(lines):
-        if line and not line.isspace():
-            return i
-    return len(lines)
-
-
-def transform_cell(code, indent_only=False):
-    """Transform IPython code to Python code."""
-    number_empty_lines = count_leading_empty_lines(code)
-    if indent_only:
-        if not code.endswith('\n'):
-            code += '\n'  # Ensure the cell has a trailing newline
-        lines = code.splitlines(keepends=True)
-        lines = leading_indent(leading_empty_lines(lines))
-        code = ''.join(lines)
-    else:
-        tm = TransformerManager()
-        code = tm.transform_cell(code)
-    return '\n' * number_empty_lines + code
-
-
-def exec_code(code, filename, ns_globals, ns_locals=None, post_mortem=False,
-              exec_fun=None, capture_last_expression=False,
-              global_warning=False):
-    """Execute code and display any exception."""
-    # Tell IPython to hide this frame (>7.16)
-    __tracebackhide__ = True
-    global SHOW_INVALID_SYNTAX_MSG
-    global SHOW_GLOBAL_MSG
-
-    if exec_fun is None:
-        exec_fun = exec
-
-    ipython_shell = get_ipython()
-    is_ipython = os.path.splitext(filename)[1] == '.ipy'
-    try:
-        if not is_ipython:
-            # TODO: remove the try-except and let the SyntaxError raise
-            # Because there should not be ipython code in a python file
-            try:
-                ast_code = ast.parse(transform_cell(code, indent_only=True))
-            except SyntaxError as e:
-                try:
-                    ast_code = ast.parse(transform_cell(code))
-                except SyntaxError:
-                    raise e from None
-                else:
-                    if SHOW_INVALID_SYNTAX_MSG:
-                        print(
-                            "\nWARNING: This is not valid Python code. "
-                            "If you want to use IPython magics, "
-                            "flexible indentation, and prompt removal, "
-                            "we recommend that you save this file with the "
-                            ".ipy extension.\n")
-                        SHOW_INVALID_SYNTAX_MSG = False
-        else:
-            ast_code = ast.parse(transform_cell(code))
-
-        # Print warning for global
-        if global_warning and SHOW_GLOBAL_MSG:
-            has_global = any(
-                isinstance(node, ast.Global) for node in ast.walk(ast_code))
-            if has_global:
-                print(
-                    "\nWARNING: This file contains a global statement, "
-                    "but it is run in an empty namespace. "
-                    "Consider using the "
-                    "'Run in console's namespace instead of an empty one' "
-                    "option, that you can find in the menu 'Run > "
-                    "Configuration per file', if you want to capture the "
-                    "namespace.\n"
-                )
-                SHOW_GLOBAL_MSG = False
-
-        if code.rstrip()[-1] == ";":
-            # Supress output with ;
-            capture_last_expression = False
-
-        if capture_last_expression:
-            ast_code, capture_last_expression = capture_last_Expr(
-                ast_code, "_spyder_out")
-
-        exec_fun(compile(ast_code, filename, 'exec'), ns_globals, ns_locals)
-
-        if capture_last_expression:
-            out = ns_globals.pop("_spyder_out", None)
-            if out is not None:
-                return out
-
-    except SystemExit as status:
-        # ignore exit(0)
-        if status.code:
-            ipython_shell.showtraceback(exception_only=True)
-    except BaseException as error:
-        if (isinstance(error, bdb.BdbQuit)
-                and ipython_shell.pdb_session):
-            # Ignore BdbQuit if we are debugging, as it is expected.
-            pass
-        elif post_mortem and isinstance(error, Exception):
-            error_type, error, tb = sys.exc_info()
-            post_mortem_excepthook(error_type, error, tb)
-        else:
-            # We ignore the call to exec
-            ipython_shell.showtraceback(tb_offset=1)
-    finally:
-        __tracebackhide__ = "__pdb_exit__"
-
-
-def get_file_code(filename, save_all=True, raise_exception=False):
-    """Retrieve the content of a file."""
-    # Get code from spyder
-    try:
-        return frontend_request(blocking=True).get_file_code(
-            filename, save_all=save_all)
-    except Exception:
-        # Maybe this is a local file
-        try:
-            with open(filename, 'r') as f:
-                return f.read()
-        except FileNotFoundError:
-            pass
-        if raise_exception:
-            raise
-        # Else return None
-        return None
-
-
-def runfile(filename=None, args=None, wdir=None, namespace=None,
-            post_mortem=False, current_namespace=False):
-    """
-    Run filename
-    args: command line arguments (string)
-    wdir: working directory
-    namespace: namespace for execution
-    post_mortem: boolean, whether to enter post-mortem mode on error
-    current_namespace: if true, run the file in the current namespace
-    """
-    return _exec_file(
-        filename, args, wdir, namespace,
-        post_mortem, current_namespace, stack_depth=1)
-
-
-def _exec_file(filename=None, args=None, wdir=None, namespace=None,
-               post_mortem=False, current_namespace=False, stack_depth=0,
-               exec_fun=None, canonic_filename=None):
-    # Tell IPython to hide this frame (>7.16)
-    __tracebackhide__ = True
-    ipython_shell = get_ipython()
-    if filename is None:
-        filename = get_current_file_name()
-        if filename is None:
-            return
-
-    if __umr__.enabled:
-        __umr__.run()
-    if args is not None and not isinstance(args, str):
-        raise TypeError("expected a character buffer object")
-
-    try:
-        file_code = get_file_code(filename, raise_exception=True)
-    except Exception:
-        print(
-            "This command failed to be executed because an error occurred"
-            " while trying to get the file code from Spyder's"
-            " editor. The error was:\n\n")
-        get_ipython().showtraceback(exception_only=True)
-        return
-
-    # Here the remote filename has been used. It must now be valid locally.
-    if canonic_filename is not None:
-        filename = canonic_filename
-    else:
-        filename = canonic(filename)
-
-    with NamespaceManager(filename, namespace, current_namespace,
-                          file_code=file_code, stack_depth=stack_depth + 1
-                          ) as (ns_globals, ns_locals):
-        sys.argv = [filename]
-        if args is not None:
-            for arg in shlex.split(args):
-                sys.argv.append(arg)
-
-        if "multiprocessing" in sys.modules:
-            # See https://github.com/spyder-ide/spyder/issues/16696
-            try:
-                sys.modules['__mp_main__'] = sys.modules['__main__']
-            except Exception:
-                pass
-
-        if wdir is not None:
-            if os.path.isdir(wdir):
-                os.chdir(wdir)
-                # See https://github.com/spyder-ide/spyder/issues/13632
-                if "multiprocessing.process" in sys.modules:
-                    try:
-                        import multiprocessing.process
-                        multiprocessing.process.ORIGINAL_DIR = os.path.abspath(
-                            wdir)
-                    except Exception:
-                        pass
-            else:
-                print("Working directory {} doesn't exist.\n".format(wdir))
-
-        try:
-            if __umr__.has_cython:
-                # Cython files
-                with io.open(filename, encoding='utf-8') as f:
-                    ipython_shell.run_cell_magic('cython', '', f.read())
-            else:
-                exec_code(file_code, filename, ns_globals, ns_locals,
-                          post_mortem=post_mortem, exec_fun=exec_fun,
-                          capture_last_expression=False,
-                          global_warning=not current_namespace)
-        finally:
-            sys.argv = ['']
-
-
-# IPykernel 6.3.0+ shadows our runfile because it depends on the Pydev
-# debugger, which adds its own runfile to builtins. So we replace it with
-# our own using exec_lines in start.py
-builtins.spyder_runfile = runfile
-
-
-def debugfile(filename=None, args=None, wdir=None, post_mortem=False,
-              current_namespace=False):
-    """
-    Debug filename
-    args: command line arguments (string)
-    wdir: working directory
-    post_mortem: boolean, included for compatiblity with runfile
-    """
-    # Tell IPython to hide this frame (>7.16)
-    __tracebackhide__ = True
-    if filename is None:
-        filename = get_current_file_name()
-        if filename is None:
-            return
-
-    shell = get_ipython()
-    if shell.is_debugging():
-        # Recursive
-        code = (
-            "runfile({}".format(repr(filename)) +
-            ", args=%r, wdir=%r, current_namespace=%r)" % (
-                args, wdir, current_namespace)
-        )
-
-        shell.pdb_session.enter_recursive_debugger(
-            code, filename, True,
-        )
-    else:
-        debugger = get_new_debugger(filename, True)
-        _exec_file(
-            filename=filename,
-            canonic_filename=debugger.canonic(filename),
-            args=args,
-            wdir=wdir,
-            current_namespace=current_namespace,
-            exec_fun=debugger.run,
-            stack_depth=1,
-        )
-
-
-builtins.debugfile = debugfile
-
-
-def runcell(cellname, filename=None, post_mortem=False):
-    """
-    Run a code cell from an editor as a file.
-
-    Parameters
-    ----------
-    cellname : str or int
-        Cell name or index.
-    filename : str
-        Needed to allow for proper traceback links.
-    post_mortem: bool
-        Automatically enter post mortem on exception.
-    """
-    # Tell IPython to hide this frame (>7.16)
-    __tracebackhide__ = True
-    return _exec_cell(cellname, filename, post_mortem, stack_depth=1)
-
-
-def _exec_cell(cellname, filename=None, post_mortem=False, stack_depth=0,
-               exec_fun=None, canonic_filename=None):
-    """
-    Execute a code cell with a given exec function.
-    """
-    # Tell IPython to hide this frame (>7.16)
-    __tracebackhide__ = True
-    if filename is None:
-        filename = get_current_file_name()
-        if filename is None:
-            return
-    ipython_shell = get_ipython()
-    try:
-        # Get code from spyder
-        cell_code = frontend_request(
-            blocking=True).run_cell(cellname, filename)
-    except Exception:
-        print("This command failed to be executed because an error occurred"
-              " while trying to get the cell code from Spyder's"
-              " editor. The error was:\n\n")
-        get_ipython().showtraceback(exception_only=True)
-        return
-
-    if not cell_code or cell_code.strip() == '':
-        print("Nothing to execute, this cell is empty.\n")
-        return
-
-    # Trigger `post_execute` to exit the additional pre-execution.
-    # See Spyder PR #7310.
-    ipython_shell.events.trigger('post_execute')
-    file_code = get_file_code(filename, save_all=False)
-
-    # Here the remote filename has been used. It must now be valid locally.
-    if canonic_filename is not None:
-        filename = canonic_filename
-    else:
-        # Normalise the filename
-        filename = canonic(filename)
-
-    with NamespaceManager(filename, current_namespace=True,
-                          file_code=file_code, stack_depth=stack_depth + 1
-                          ) as (ns_globals, ns_locals):
-        return exec_code(cell_code, filename, ns_globals, ns_locals,
-                         post_mortem=post_mortem, exec_fun=exec_fun,
-                         capture_last_expression=True)
-
-
-builtins.runcell = runcell
-
-
-def debugcell(cellname, filename=None, post_mortem=False):
-    """Debug a cell."""
-    # Tell IPython to hide this frame (>7.16)
-    __tracebackhide__ = True
-    if filename is None:
-        filename = get_current_file_name()
-        if filename is None:
-            return
-
-    shell = get_ipython()
-    if shell.is_debugging():
-        # Recursive
-        code = (
-            "runcell({}, ".format(repr(cellname)) +
-            "{})".format(repr(filename))
-        )
-        shell.pdb_session.enter_recursive_debugger(
-            code, filename, False,
-        )
-    else:
-        debugger = get_new_debugger(filename, False)
-        _exec_cell(
-            cellname=cellname,
-            filename=filename,
-            canonic_filename=debugger.canonic(filename),
-            exec_fun=debugger.run,
-            stack_depth=1
-        )
-
-
-builtins.debugcell = debugcell
-
-
-def cell_count(filename=None):
-    """
-    Get the number of cells in a file.
-
-    Parameters
-    ----------
-    filename : str
-        The file to get the cells from. If None, the currently opened file.
-    """
-    if filename is None:
-        filename = get_current_file_name()
-        if filename is None:
-            raise RuntimeError('Could not get cell count from frontend.')
-    try:
-        # Get code from spyder
-        cell_count = frontend_request(blocking=True).cell_count(filename)
-        return cell_count
-    except Exception:
-        etype, error, tb = sys.exc_info()
-        raise etype(error)
-
-
-builtins.cell_count = cell_count
-
 
 # =============================================================================
 # PYTHONPATH and sys.path Adjustments
