@@ -16,7 +16,7 @@ def test_user_sitepackages_in_pathlist():
     """Test that we include user site-packages in pathlist."""
     if sys.platform.startswith('linux'):
         user_path = 'local'
-    elif sys.platform == 'darwin':
+    elif (sys.platform == 'darwin' or sys.platform.startswith('freebsd')):
         user_path = os.path.expanduser('~/.local')
     else:
         user_path = 'Roaming'
