@@ -6,6 +6,7 @@
 """Utility functions."""
 
 import ast
+import builtins
 import os
 import re
 import sys
@@ -96,7 +97,7 @@ def path_is_library(path, initial_pathlist=None):
         return False
 
 
-def capture_last_Expr(code_ast, out_varname):
+def capture_last_Expr(code_ast, out_varname, global_ns):
     """Parse line and modify code to capture in globals the last expression."""
     # Modify ast code to capture the last expression
     capture_last_expression = False
@@ -104,11 +105,13 @@ def capture_last_Expr(code_ast, out_varname):
         len(code_ast.body)
         and isinstance(code_ast.body[-1], ast.Expr)
     ):
+        global_ns["__spyder_builtins__"] = builtins
         capture_last_expression = True
         expr_node = code_ast.body[-1]
         # Create new assign node
         assign_node = ast.parse(
-            'globals()[{}] = None'.format(repr(out_varname))).body[0]
+            '__spyder_builtins__.globals()[{}] = None'.format(
+                repr(out_varname))).body[0]
         # Replace None by the value
         assign_node.value = expr_node.value
         # Fix line number and column offset
@@ -118,12 +121,100 @@ def capture_last_Expr(code_ast, out_varname):
             # Exists from 3.8, necessary from 3.11
             assign_node.end_lineno = expr_node.end_lineno
             if assign_node.lineno == assign_node.end_lineno:
-                # Add 'globals()[{}] = ' and remove 'None'
+                # Add '__spyder_builtins__.globals()[{}] = ' and remove 'None'
                 assign_node.end_col_offset += expr_node.end_col_offset - 4
             else:
                 assign_node.end_col_offset = expr_node.end_col_offset
         code_ast.body[-1] = assign_node
     return code_ast, capture_last_expression
+
+
+def exec_encapsulate_locals(
+    code_ast, globals, locals, exec_fun=None, filename=None
+):
+    """
+    Execute by encapsulating locals if needed.
+
+    Notes
+    ----- 
+    * In general, the dict returned by locals() might or might not be modified.
+      In this case, the encapsulated dict can not.
+    """
+    use_locals_hack = locals is not None and locals is not globals
+    if use_locals_hack:
+        globals["__spyder_builtins__"] = builtins
+
+        # Mitigates a behaviour of CPython that makes it difficult
+        # to work with exec and the local namespace
+        # See:
+        #  - https://bugs.python.org/issue41918
+        #  - https://bugs.python.org/issue46153
+        #  - https://bugs.python.org/issue21161
+        #  - spyder-ide/spyder#13909
+        #  - spyder-ide/spyder-kernels#345
+        #
+        # The idea here is that the best way to emulate being in a
+        # function is to actually execute the code in a function.
+        # A function called `_spyderpdb_code` is created and
+        # called. It will first load the locals, execute the code,
+        # and then update the locals.
+        #
+        # One limitation of this approach is that locals() is only
+        # a copy of the curframe locals. This means that closures
+        # for example are early binding instead of late binding.
+
+        # Create a function
+        indent = "    "
+        code = ["def _spyderpdb_code():"]
+
+        # Add locals in globals
+        # If the debugger is recursive, the globals could already
+        # have a _spyderpdb_locals as it might be shared between
+        # levels
+        if "_spyderpdb_locals" in globals:
+            globals["_spyderpdb_locals"].append(locals)
+        else:
+            globals["_spyderpdb_locals"] = [locals]
+
+        # Load locals if they have a valid name
+        # In comprehensions, locals could contain ".0" for example
+        code += [indent + "{k} = _spyderpdb_locals[-1]['{k}']".format(
+            k=k) for k in locals if k.isidentifier()]
+
+        # The code comes here
+
+        # Update the locals
+        code += [indent + "_spyderpdb_locals[-1].update("
+                 "__spyder_builtins__.locals())"]
+
+        # Run the function
+        code += ["_spyderpdb_code()"]
+
+        # Parse the function
+        fun_ast = ast.parse('\n'.join(code) + '\n')
+
+        # Inject code_ast in the function before the locals update
+        fun_ast.body[0].body = (
+            fun_ast.body[0].body[:-1]  # The locals
+            + code_ast.body  # Code to run
+            + fun_ast.body[0].body[-1:]  # Locals update
+        )
+        code_ast = fun_ast
+
+    try:
+        if exec_fun is None:
+            exec_fun = exec
+        if filename is None:
+            filename = "<stdin>"
+        exec_fun(compile(code_ast, filename, "exec"), globals)
+    finally:
+        if use_locals_hack:
+            # Cleanup code
+            globals.pop("_spyderpdb_code", None)
+            if len(globals["_spyderpdb_locals"]) > 1:
+                del globals["_spyderpdb_locals"][-1]
+            else:
+                del globals["_spyderpdb_locals"]
 
 
 def canonic(filename):
