@@ -7,9 +7,6 @@
 # -----------------------------------------------------------------------------
 
 """Utilities and wrappers around inspect module"""
-
-from __future__ import print_function
-
 import builtins
 import inspect
 import re
@@ -110,11 +107,16 @@ def getdoc(obj):
             doc['note'] = 'Function'
         doc['name'] = obj.__name__
         if inspect.isfunction(obj):
-            (args, varargs, varkw, defaults, kwonlyargs, kwonlydefaults,
-             annotations) = inspect.getfullargspec(obj)
-            doc['argspec'] = inspect.formatargspec(
-                args, varargs, varkw, defaults, kwonlyargs, kwonlydefaults,
-                annotations, formatvalue=lambda o: '='+repr(o))
+            # This is necessary to catch errors for objects without a
+            # signature, like numpy.where.
+            # Fixes spyder-ide/spyder#21148
+            try:
+                sig = inspect.signature(obj)
+            except ValueError:
+                sig = getargspecfromtext(doc['docstring'])
+                if not sig:
+                    sig = '(...)'
+            doc['argspec'] = str(sig)
             if name == '<lambda>':
                 doc['name'] = name + ' lambda '
                 doc['argspec'] = doc['argspec'][1:-1] # remove parentheses
@@ -164,51 +166,78 @@ def getsource(obj):
 
 
 def getsignaturefromtext(text, objname):
-    """Get object signatures from text (object documentation)
-    Return a list containing a single string in most cases
-    Example of multiple signatures: PyQt5 objects"""
+    """Get object signature from text (i.e. object documentation)."""
     if isinstance(text, dict):
         text = text.get('docstring', '')
+
     # Regexps
-    oneline_re = objname + r'\([^\)].+?(?<=[\w\]\}\'"])\)(?!,)'
-    multiline_re = objname + r'\([^\)]+(?<=[\w\]\}\'"])\)(?!,)'
-    multiline_end_parenleft_re = r'(%s\([^\)]+(\),\n.+)+(?<=[\w\]\}\'"])\))'
+    args_re = r'(\(.+?\))'
+    if objname:
+        signature_re = objname + args_re
+    else:
+        identifier_re = r'(\w+)'
+        signature_re = identifier_re + args_re
+
     # Grabbing signatures
     if not text:
         text = ''
-    sigs_1 = re.findall(oneline_re + '|' + multiline_re, text)
-    sigs_2 = [g[0] for g in re.findall(multiline_end_parenleft_re % objname, text)]
-    all_sigs = sigs_1 + sigs_2
-    # The most relevant signature is usually the first one. There could be
-    # others in doctests but those are not so important
-    if all_sigs:
-        return all_sigs[0]
-    else:
-        return ''
 
-# Fix for Issue 1953
-# TODO: Add more signatures and remove this hack in 2.4
-getsignaturesfromtext = getsignaturefromtext
+    sigs = re.findall(signature_re, text)
+
+    # The most relevant signature is usually the first one. There could be
+    # others in doctests or other places, but those are not so important.
+    sig = ''
+    if sigs:
+        # Default signatures returned by IPython.
+        # Notes:
+        # * These are not real signatures but only used to provide a
+        #   placeholder.
+        # * We skip them if we can find other signatures in `text`.
+        # * This is necessary because we also use this function in Spyder
+        #   to parse the content of inspect replies that come from the
+        #   kernel, which can include these signatures.
+        default_ipy_sigs = ['(*args, **kwargs)', '(self, /, *args, **kwargs)']
+
+        if objname:
+            real_sigs = [s for s in sigs if s not in default_ipy_sigs]
+
+            if real_sigs:
+                sig = real_sigs[0]
+            else:
+                sig = sigs[0]
+        else:
+            valid_sigs = [s for s in sigs if s[0].isidentifier()]
+
+            if valid_sigs:
+                real_sigs = [
+                    s for s in valid_sigs if s[1] not in default_ipy_sigs
+                ]
+
+                if real_sigs:
+                    sig = real_sigs[0][1]
+                else:
+                    sig = valid_sigs[0][1]
+
+    return sig
 
 
 def getargspecfromtext(text):
     """
     Try to get the formatted argspec of a callable from the first block of its
-    docstring
+    docstring.
     
-    This will return something like
-    '(foo, bar, k=1)'
+    This will return something like `(x, y, k=1)`.
     """
     blocks = text.split("\n\n")
-    first_block = blocks[0].strip()
+    first_block = blocks[0].strip().replace('\n', '')
     return getsignaturefromtext(first_block, '')
 
 
 def getargsfromtext(text, objname):
-    """Get arguments from text (object documentation)"""
+    """Get arguments from text (object documentation)."""
     signature = getsignaturefromtext(text, objname)
     if signature:
-        argtxt = signature[signature.find('(')+1:-1]
+        argtxt = signature[signature.find('(') + 1:-1]
         return argtxt.split(',')
 
 
@@ -323,20 +352,3 @@ def isdefined(obj, force_import=False, namespace=None):
                 return False
         base += '.'+attr
     return True
-    
-
-if __name__ == "__main__":
-    class Test:
-        def method(self, x, y=2):
-            pass
-    print(getargtxt(Test.__init__))  # spyder: test-skip
-    print(getargtxt(Test.method))  # spyder: test-skip
-    print(isdefined('numpy.take', force_import=True))  # spyder: test-skip
-    print(isdefined('__import__'))  # spyder: test-skip
-    print(isdefined('.keys', force_import=True))  # spyder: test-skip
-    print(getobj('globals'))  # spyder: test-skip
-    print(getobj('globals().keys'))  # spyder: test-skip
-    print(getobj('+scipy.signal.'))  # spyder: test-skip
-    print(getobj('4.'))  # spyder: test-skip
-    print(getdoc(sorted))  # spyder: test-skip
-    print(getargtxt(sorted))  # spyder: test-skip
