@@ -11,13 +11,6 @@
 from spyder_kernels.utils.misc import is_module_installed
 
 
-# Mapping of inline figure formats
-INLINE_FIGURE_FORMATS = {
-    '0': 'png',
-    '1': 'svg'
-}
-
-
 # Inline backend
 if is_module_installed('matplotlib_inline'):
     inline_backend = 'module://matplotlib_inline.backend_inline'
@@ -27,30 +20,21 @@ else:
 
 # Mapping of matlotlib backends options to Spyder
 MPL_BACKENDS_TO_SPYDER = {
-    inline_backend: 0,
-    'Qt5Agg': 2,
-    'QtAgg': 2,  # For Matplotlib 3.5+
-    'TkAgg': 3,
-    'MacOSX': 4,
+    'inline': 'inline',  # For Matplotlib >=3.9
+    inline_backend: "inline",  # For Matplotlib <3.9
+    'qt5agg': 'qt',
+    'qtagg': 'qt',  # For Matplotlib 3.5+
+    'tkagg': 'tk',
+    'macosx': 'osx',
 }
 
 
 def automatic_backend():
     """Get Matplolib automatic backend option."""
     if is_module_installed('PyQt5'):
-        auto_backend = 'qt5'
+        auto_backend = 'qt'
     elif is_module_installed('_tkinter'):
         auto_backend = 'tk'
     else:
         auto_backend = 'inline'
     return auto_backend
-
-
-# Mapping of Spyder options to backends
-MPL_BACKENDS_FROM_SPYDER = {
-    '0': 'inline',
-    '1': automatic_backend(),
-    '2': 'qt5',
-    '3': 'tk',
-    '4': 'osx'
-}
