@@ -9,9 +9,6 @@
 """
 Utilities to build a namespace view.
 """
-
-from __future__ import print_function
-
 from itertools import islice
 import inspect
 import re
@@ -592,7 +589,8 @@ def is_callable_or_module(value):
 def globalsfilter(input_dict, check_all=False, filters=None,
                   exclude_private=None, exclude_capitalized=None,
                   exclude_uppercase=None, exclude_unsupported=None,
-                  excluded_names=None, exclude_callables_and_modules=None):
+                  excluded_names=None, exclude_callables_and_modules=None,
+                  filter_on=True):
     """Keep objects in namespace view according to different criteria."""
     output_dict = {}
     def _is_string(obj):
@@ -608,7 +606,7 @@ def globalsfilter(input_dict, check_all=False, filters=None,
             (exclude_callables_and_modules and is_callable_or_module(value)) or
             (exclude_unsupported and
              not is_supported(value, check_all=check_all, filters=filters))
-        )
+        ) and filter_on
         if not excluded:
             output_dict[key] = value
     return output_dict
@@ -620,7 +618,8 @@ def globalsfilter(input_dict, check_all=False, filters=None,
 REMOTE_SETTINGS = ('check_all', 'exclude_private', 'exclude_uppercase',
                    'exclude_capitalized', 'exclude_unsupported',
                    'excluded_names', 'minmax', 'show_callable_attributes',
-                   'show_special_attributes', 'exclude_callables_and_modules')
+                   'show_special_attributes', 'exclude_callables_and_modules',
+                   'filter_on')
 
 
 def get_supported_types():
@@ -676,7 +675,7 @@ def get_remote_data(data, settings, mode, more_excluded_names=None):
         exclude_capitalized=settings['exclude_capitalized'],
         exclude_unsupported=settings['exclude_unsupported'],
         exclude_callables_and_modules=settings['exclude_callables_and_modules'],
-        excluded_names=excluded_names)
+        excluded_names=excluded_names, filter_on=settings['filter_on'])
 
 
 def make_remote_view(data, settings, more_excluded_names=None):
