@@ -16,6 +16,7 @@ import os
 import copy
 
 # Third party imports
+from PIL import ImageFile
 import pytest
 import numpy as np
 
@@ -24,8 +25,9 @@ import spyder_kernels.utils.iofuncs as iofuncs
 
 
 # Full path to this file's parent directory for loading data
-LOCATION = os.path.realpath(os.path.join(os.getcwd(),
-                                         os.path.dirname(__file__)))
+LOCATION = os.path.realpath(
+    os.path.join(os.getcwd(), os.path.dirname(__file__))
+)
 
 
 # =============================================================================
@@ -87,7 +89,7 @@ def spydata_values():
     B = 'ham'
     C = np.eye(3)
     D = {'a': True, 'b': np.eye(4, dtype=np.complex128)}
-    E = [np.eye(2, dtype=np.int64), 42.0, np.eye(3, dtype=np.bool_)]
+    E = [np.eye(2, dtype=np.int64), 42.0, np.eye(3, dtype=np.bool_), np.eye(4, dtype=object)]
     return {'A': A, 'B': B, 'C': C, 'D': D, 'E': E}
 
 
@@ -336,6 +338,27 @@ def test_spydata_export(input_namespace, expected_namespace,
                 os.remove(path)
             except (IOError, OSError, PermissionError):
                 pass
+
+
+def test_save_load_hdf5_files(tmp_path):
+    """Simple test to check that we can save and load HDF5 files."""
+    h5_file = tmp_path / "test.h5"
+    data = {'a' : [1, 2, 3, 4], 'b' : 4.5}
+    iofuncs.save_hdf5(data, h5_file)
+
+    expected = ({'a': np.array([1, 2, 3, 4]), 'b': np.array(4.5)}, None)
+    assert repr(iofuncs.load_hdf5(h5_file)) == repr(expected)
+
+
+def test_load_dicom_files():
+    """Check that we can load DICOM files."""
+    # This test pass locally but we need to set the variable below for it to
+    # pass on CIs.
+    # See https://stackoverflow.com/a/47958486/438386 for context.
+    ImageFile.LOAD_TRUNCATED_IMAGES = True
+
+    data = iofuncs.load_dicom(os.path.join(LOCATION, 'data.dcm'))
+    assert data[0]['data'].shape == (512, 512)
 
 
 if __name__ == "__main__":
